utils: Move Labels to a separate file
This will be needed to avoid a circular dependency when moving some
Flatpak-handling logic from api.py to repo_utils.py.

We import Labels in __init__.py to keep compatibility with external
users of the utils.Labels API (atomic-client in particular.)
